% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_usa_names}
\alias{clean_usa_names}
\title{Standardize USA State Names}
\usage{
clean_usa_names(country_vec)
}
\arguments{
\item{country_vec}{A character vector of country or state names}
}
\value{
A character vector where state names are replaced with "USA"
}
\description{
Standardize USA State Names
}
\examples{
clean_usa_names(c("Mexico", "California", "France", "Texas"))
}
