% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location Name column}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{A character vector of location names}
}
\value{
A cleaned character vector in title case without country prefixes
}
\description{
Clean Location Name column
}
