% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{eq_map}
\alias{eq_map}
\title{Interactive Earthquake Map}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{A filtered data frame containing earthquake data.}

\item{annot_col}{A character string specifying the column name to use for pop-up annotations.}
}
\value{
A leaflet object displaying the map.
}
\description{
Interactive Earthquake Map
}
\examples{
\dontrun{
eq_map(sample_data, annot_col = "DATE")
}
}
