% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Read and Filter NOAA Earthquake Data}
\usage{
eq_read_data(file_path)
}
\arguments{
\item{file_path}{A character string pointing to the .tsv file.}
}
\value{
A tibble with the raw data, skipping the search parameter header.
}
\description{
Read and Filter NOAA Earthquake Data
}
